+++ cache-service-test/mvn_dependency_tree.txt
@@ -47,10 +47,7 @@ uk.gov.gchq.palisade:cache-service-test:jar:0.4.0-SNAPSHOT
 |  +- io.fabric8:kubernetes-model-common:jar:4.3.0:compile
 |  +- com.fasterxml.jackson.module:jackson-module-jaxb-annotations:jar:2.9.9:compile
 |  +- javax.validation:validation-api:jar:2.0.1.Final:compile
-|  +- org.glassfish:javax.el:jar:3.0.0:compile
-|  +- javax.annotation:javax.annotation-api:jar:1.3.2:compile
-|  \- javax.xml.bind:jaxb-api:jar:2.3.1:compile
-|     \- javax.activation:javax.activation-api:jar:1.2.0:compile
+|  \- org.glassfish:javax.el:jar:3.0.0:compile
 +- org.springframework.boot:spring-boot-starter:jar:2.1.8.RELEASE:compile
 |  +- org.springframework.boot:spring-boot:jar:2.1.8.RELEASE:compile
 |  |  \- org.springframework:spring-context:jar:5.1.9.RELEASE:compile
@@ -60,6 +57,7 @@ uk.gov.gchq.palisade:cache-service-test:jar:0.4.0-SNAPSHOT
 |  |  |  \- ch.qos.logback:logback-core:jar:1.2.3:compile
 |  |  \- org.apache.logging.log4j:log4j-to-slf4j:jar:2.11.2:compile
 |  |     \- org.apache.logging.log4j:log4j-api:jar:2.11.2:compile
+|  +- javax.annotation:javax.annotation-api:jar:1.3.2:compile
 |  +- org.springframework:spring-core:jar:5.1.9.RELEASE:compile
 |  |  \- org.springframework:spring-jcl:jar:5.1.9.RELEASE:compile
 |  \- org.yaml:snakeyaml:jar:1.23:compile
@@ -101,6 +99,8 @@ uk.gov.gchq.palisade:cache-service-test:jar:0.4.0-SNAPSHOT
 |  +- com.github.spotbugs:spotbugs-annotations:jar:3.1.10:test
 |  \- org.testcontainers:testcontainers:jar:1.10.5:test
 |     +- org.jetbrains:annotations:jar:16.0.3:test
+|     +- javax.xml.bind:jaxb-api:jar:2.3.1:test
+|     |  \- javax.activation:javax.activation-api:jar:1.2.0:test
 |     +- org.rnorth.duct-tape:duct-tape:jar:1.0.7:test
 |     +- org.rnorth.visible-assertions:visible-assertions:jar:2.1.2:test
 |     |  \- net.java.dev.jna:jna:jar:4.5.2:test
