+++ palisade-service-test/mvn_dependency_tree.txt
@@ -34,9 +34,12 @@ uk.gov.gchq.palisade:palisade-service-test:jar:0.4.0-SNAPSHOT
 |  |  +- org.apache.httpcomponents:httpclient:jar:4.5.9:runtime
 |  |  |  +- org.apache.httpcomponents:httpcore:jar:4.4.12:runtime
 |  |  |  \- commons-codec:commons-codec:jar:1.11:runtime
-|  |  \- com.google.inject:guice:jar:4.1.0:runtime
-|  |     +- javax.inject:javax.inject:jar:1:runtime
-|  |     \- aopalliance:aopalliance:jar:1.0:runtime
+|  |  +- com.google.inject:guice:jar:4.1.0:runtime
+|  |  |  +- javax.inject:javax.inject:jar:1:runtime
+|  |  |  \- aopalliance:aopalliance:jar:1.0:runtime
+|  |  +- com.fasterxml.jackson.core:jackson-annotations:jar:2.9.0:compile
+|  |  +- com.fasterxml.jackson.core:jackson-core:jar:2.9.9:compile
+|  |  \- com.fasterxml.jackson.core:jackson-databind:jar:2.9.9.3:compile
 |  +- io.etcd:jetcd-core:jar:0.3.0:compile
 |  |  +- io.etcd:jetcd-common:jar:0.3.0:compile
 |  |  +- io.etcd:jetcd-resolver:jar:0.3.0:compile
@@ -60,8 +63,9 @@ uk.gov.gchq.palisade:palisade-service-test:jar:0.4.0-SNAPSHOT
 |  |  |  +- com.google.api.grpc:proto-google-common-protos:jar:1.16.0:compile
 |  |  |  \- io.grpc:grpc-protobuf-lite:jar:1.17.1:compile
 |  |  +- io.grpc:grpc-stub:jar:1.17.1:compile
-|  |  \- io.grpc:grpc-grpclb:jar:1.17.1:compile
-|  |     \- com.google.protobuf:protobuf-java-util:jar:3.5.1:compile
+|  |  +- io.grpc:grpc-grpclb:jar:1.17.1:compile
+|  |  |  \- com.google.protobuf:protobuf-java-util:jar:3.5.1:compile
+|  |  \- org.slf4j:slf4j-api:jar:1.7.28:compile
 |  +- io.fabric8:kubernetes-client:jar:4.1.0:compile
 |  |  +- com.squareup.okhttp3:okhttp:jar:3.8.1:compile
 |  |  |  \- com.squareup.okio:okio:jar:1.13.0:compile
@@ -98,21 +102,32 @@ uk.gov.gchq.palisade:palisade-service-test:jar:0.4.0-SNAPSHOT
 |  |  +- io.github.openfeign:feign-slf4j:jar:10.2.3:compile
 |  |  \- io.github.openfeign:feign-hystrix:jar:10.2.3:compile
 |  |     \- com.netflix.hystrix:hystrix-core:jar:1.5.18:compile
-|  \- org.springframework.cloud:spring-cloud-starter-ribbon:jar:1.4.7.RELEASE:compile
-|     \- org.springframework.cloud:spring-cloud-starter-netflix-ribbon:jar:2.1.3.RELEASE:compile
-|        +- org.springframework.cloud:spring-cloud-starter-netflix-archaius:jar:2.1.3.RELEASE:compile
-|        +- com.netflix.ribbon:ribbon:jar:2.3.0:compile
-|        |  +- com.netflix.ribbon:ribbon-transport:jar:2.3.0:runtime
-|        |  |  +- io.reactivex:rxnetty-contexts:jar:0.4.9:runtime
-|        |  |  \- io.reactivex:rxnetty-servo:jar:0.4.9:runtime
-|        |  \- io.reactivex:rxnetty:jar:0.4.9:runtime
-|        +- com.netflix.ribbon:ribbon-core:jar:2.3.0:compile
-|        +- com.netflix.ribbon:ribbon-httpclient:jar:2.3.0:compile
-|        |  +- commons-collections:commons-collections:jar:3.2.2:runtime
-|        |  \- com.netflix.netflix-commons:netflix-commons-util:jar:0.3.0:runtime
-|        +- com.netflix.ribbon:ribbon-loadbalancer:jar:2.3.0:compile
-|        |  \- com.netflix.netflix-commons:netflix-statistics:jar:0.1.1:runtime
-|        \- io.reactivex:rxjava:jar:1.3.8:compile
+|  +- org.springframework.cloud:spring-cloud-starter-ribbon:jar:1.4.7.RELEASE:compile
+|  |  \- org.springframework.cloud:spring-cloud-starter-netflix-ribbon:jar:2.1.3.RELEASE:compile
+|  |     +- org.springframework.cloud:spring-cloud-starter-netflix-archaius:jar:2.1.3.RELEASE:compile
+|  |     +- com.netflix.ribbon:ribbon:jar:2.3.0:compile
+|  |     |  +- com.netflix.ribbon:ribbon-transport:jar:2.3.0:runtime
+|  |     |  |  +- io.reactivex:rxnetty-contexts:jar:0.4.9:runtime
+|  |     |  |  \- io.reactivex:rxnetty-servo:jar:0.4.9:runtime
+|  |     |  \- io.reactivex:rxnetty:jar:0.4.9:runtime
+|  |     +- com.netflix.ribbon:ribbon-core:jar:2.3.0:compile
+|  |     +- com.netflix.ribbon:ribbon-httpclient:jar:2.3.0:compile
+|  |     |  +- commons-collections:commons-collections:jar:3.2.2:runtime
+|  |     |  \- com.netflix.netflix-commons:netflix-commons-util:jar:0.3.0:runtime
+|  |     +- com.netflix.ribbon:ribbon-loadbalancer:jar:2.3.0:compile
+|  |     |  \- com.netflix.netflix-commons:netflix-statistics:jar:0.1.1:runtime
+|  |     \- io.reactivex:rxjava:jar:1.3.8:compile
+|  \- uk.gov.gchq.palisade:common:jar:0.4.0-SNAPSHOT:compile
+|     +- com.fasterxml.jackson.datatype:jackson-datatype-jdk8:jar:2.9.9:compile
+|     +- com.fasterxml.jackson.datatype:jackson-datatype-jsr310:jar:2.9.9:compile
+|     +- commons-io:commons-io:jar:2.6:compile
+|     +- org.apache.commons:commons-lang3:jar:3.8.1:compile
+|     \- org.apache.avro:avro:jar:1.8.2:compile
+|        +- org.codehaus.jackson:jackson-core-asl:jar:1.9.13:compile
+|        +- org.codehaus.jackson:jackson-mapper-asl:jar:1.9.13:compile
+|        +- com.thoughtworks.paranamer:paranamer:jar:2.7:compile
+|        +- org.xerial.snappy:snappy-java:jar:1.1.1.3:compile
+|        \- org.tukaani:xz:jar:1.5:compile
 +- org.springframework.boot:spring-boot-starter:jar:2.1.8.RELEASE:compile
 |  +- org.springframework.boot:spring-boot:jar:2.1.8.RELEASE:compile
 |  |  \- org.springframework:spring-context:jar:5.1.9.RELEASE:compile
@@ -148,21 +163,6 @@ uk.gov.gchq.palisade:palisade-service-test:jar:0.4.0-SNAPSHOT
 |     +- org.hdrhistogram:HdrHistogram:jar:2.1.9:compile
 |     \- org.latencyutils:LatencyUtils:jar:2.0.3:compile
 +- com.github.tomakehurst:wiremock-standalone:jar:2.20.0:compile
-+- uk.gov.gchq.palisade:common:jar:0.4.0-SNAPSHOT:compile
-|  +- com.fasterxml.jackson.core:jackson-core:jar:2.9.9:compile
-|  +- com.fasterxml.jackson.core:jackson-annotations:jar:2.9.0:compile
-|  +- com.fasterxml.jackson.core:jackson-databind:jar:2.9.9.3:compile
-|  +- com.fasterxml.jackson.datatype:jackson-datatype-jdk8:jar:2.9.9:compile
-|  +- com.fasterxml.jackson.datatype:jackson-datatype-jsr310:jar:2.9.9:compile
-|  +- commons-io:commons-io:jar:2.6:compile
-|  +- org.apache.commons:commons-lang3:jar:3.8.1:compile
-|  +- org.slf4j:slf4j-api:jar:1.7.28:compile
-|  \- org.apache.avro:avro:jar:1.8.2:compile
-|     +- org.codehaus.jackson:jackson-core-asl:jar:1.9.13:compile
-|     +- org.codehaus.jackson:jackson-mapper-asl:jar:1.9.13:compile
-|     +- com.thoughtworks.paranamer:paranamer:jar:2.7:compile
-|     +- org.xerial.snappy:snappy-java:jar:1.1.1.3:compile
-|     \- org.tukaani:xz:jar:1.5:compile
 \- org.springframework.boot:spring-boot-starter-test:jar:2.1.8.RELEASE:test
    +- org.springframework.boot:spring-boot-test:jar:2.1.8.RELEASE:test
    +- org.springframework.boot:spring-boot-test-autoconfigure:jar:2.1.8.RELEASE:test
@@ -182,5 +182,3 @@ uk.gov.gchq.palisade:palisade-service-test:jar:0.4.0-SNAPSHOT
    |  \- com.vaadin.external.google:android-json:jar:0.0.20131108.vaadin1:test
    +- org.springframework:spring-test:jar:5.1.9.RELEASE:test
    \- org.xmlunit:xmlunit-core:jar:2.6.3:test
-      \- javax.xml.bind:jaxb-api:jar:2.3.1:test
-         \- javax.activation:javax.activation-api:jar:1.2.0:test
