+++ resource-service-test/mvn_dependency_tree.txt
@@ -1,5 +1,29 @@
 uk.gov.gchq.palisade:resource-service-test:jar:0.4.0-SNAPSHOT
 +- uk.gov.gchq.palisade:resource-service:jar:jar-with-dependencies:0.4.0-SNAPSHOT:compile
+|  +- com.netflix.eureka:eureka-client:jar:1.9.13:compile
+|  |  +- org.codehaus.jettison:jettison:jar:1.3.7:compile
+|  |  |  \- stax:stax-api:jar:1.0.1:compile
+|  |  +- com.netflix.netflix-commons:netflix-eventbus:jar:0.3.0:runtime
+|  |  |  +- com.netflix.netflix-commons:netflix-infix:jar:0.3.0:runtime
+|  |  |  |  +- commons-jxpath:commons-jxpath:jar:1.3:runtime
+|  |  |  |  +- joda-time:joda-time:jar:2.10.3:runtime
+|  |  |  |  \- org.antlr:antlr-runtime:jar:3.4:runtime
+|  |  |  |     +- org.antlr:stringtemplate:jar:3.2.1:runtime
+|  |  |  |     \- antlr:antlr:jar:2.7.7:runtime
+|  |  |  \- org.apache.commons:commons-math:jar:2.2:runtime
+|  |  +- com.thoughtworks.xstream:xstream:jar:1.4.11.1:runtime
+|  |  |  +- xmlpull:xmlpull:jar:1.1.3.1:runtime
+|  |  |  \- xpp3:xpp3_min:jar:1.1.4c:runtime
+|  |  +- com.netflix.archaius:archaius-core:jar:0.7.6:runtime
+|  |  |  \- commons-configuration:commons-configuration:jar:1.8:runtime
+|  |  |     \- commons-lang:commons-lang:jar:2.6:runtime
+|  |  +- javax.ws.rs:jsr311-api:jar:1.1.1:compile
+|  |  +- com.netflix.servo:servo-core:jar:0.12.21:runtime
+|  |  +- com.sun.jersey:jersey-client:jar:1.19.1:runtime
+|  |  +- com.sun.jersey.contribs:jersey-apache-client4:jar:1.19.1:runtime
+|  |  \- com.google.inject:guice:jar:4.1.0:runtime
+|  |     +- javax.inject:javax.inject:jar:1:runtime
+|  |     \- aopalliance:aopalliance:jar:1.0:runtime
 |  +- io.etcd:jetcd-core:jar:0.3.0:compile
 |  |  +- io.etcd:jetcd-common:jar:0.3.0:compile
 |  |  +- io.etcd:jetcd-resolver:jar:0.3.0:compile
@@ -73,11 +97,10 @@ uk.gov.gchq.palisade:resource-service-test:jar:0.4.0-SNAPSHOT
 |  |     \- org.springframework:spring-expression:jar:5.1.9.RELEASE:compile
 |  +- org.springframework.cloud:spring-cloud-commons:jar:2.1.3.RELEASE:compile
 |  |  \- org.springframework.security:spring-security-crypto:jar:5.1.6.RELEASE:compile
-|  +- org.springframework.cloud:spring-cloud-netflix-eureka-client:jar:2.1.3.RELEASE:compile
-|  |  \- org.springframework.cloud:spring-cloud-netflix-hystrix:jar:2.1.3.RELEASE:compile
-|  |     \- org.springframework.boot:spring-boot-starter-aop:jar:2.1.8.RELEASE:compile
-|  |        \- org.aspectj:aspectjweaver:jar:1.9.4:compile
-|  \- uk.gov.gchq.palisade:hadoop-resource:jar:0.4.0-SNAPSHOT:compile
+|  \- org.springframework.cloud:spring-cloud-netflix-eureka-client:jar:2.1.3.RELEASE:compile
+|     \- org.springframework.cloud:spring-cloud-netflix-hystrix:jar:2.1.3.RELEASE:compile
+|        \- org.springframework.boot:spring-boot-starter-aop:jar:2.1.8.RELEASE:compile
+|           \- org.aspectj:aspectjweaver:jar:1.9.4:compile
 +- com.google.guava:guava:jar:28.1-jre:test
 |  +- com.google.guava:failureaccess:jar:1.0.1:test
 |  +- com.google.guava:listenablefuture:jar:9999.0-empty-to-avoid-conflict-with-guava:test
@@ -89,6 +112,7 @@ uk.gov.gchq.palisade:resource-service-test:jar:0.4.0-SNAPSHOT
 +- org.apache.commons:commons-lang3:jar:3.8.1:test
 +- org.apache.hadoop:hadoop-common:jar:3.2.0:compile
 |  +- org.apache.hadoop:hadoop-annotations:jar:3.2.0:compile
+|  |  \- jdk.tools:jdk.tools:jar:1.8:system
 |  +- commons-cli:commons-cli:jar:1.2:compile
 |  +- org.apache.commons:commons-math3:jar:3.1.1:compile
 |  +- org.apache.httpcomponents:httpclient:jar:4.5.9:compile
@@ -108,10 +132,8 @@ uk.gov.gchq.palisade:resource-service-test:jar:0.4.0-SNAPSHOT
 |  |  \- org.eclipse.jetty:jetty-xml:jar:9.4.19.v20190610:compile
 |  +- javax.servlet.jsp:jsp-api:jar:2.1:runtime
 |  +- com.sun.jersey:jersey-core:jar:1.19:compile
-|  |  \- javax.ws.rs:jsr311-api:jar:1.1.1:compile
 |  +- com.sun.jersey:jersey-servlet:jar:1.19:compile
 |  +- com.sun.jersey:jersey-json:jar:1.19:compile
-|  |  +- org.codehaus.jettison:jettison:jar:1.1:compile
 |  |  +- com.sun.xml.bind:jaxb-impl:jar:2.2.3-1:compile
 |  |  +- org.codehaus.jackson:jackson-core-asl:jar:1.9.2:compile
 |  |  +- org.codehaus.jackson:jackson-mapper-asl:jar:1.9.2:compile
@@ -180,11 +202,7 @@ uk.gov.gchq.palisade:resource-service-test:jar:0.4.0-SNAPSHOT
 |  |  |  +- com.zaxxer:HikariCP-java7:jar:2.4.12:test
 |  |  |  \- com.microsoft.sqlserver:mssql-jdbc:jar:6.4.0.jre8:test
 |  |  +- org.apache.hadoop:hadoop-yarn-server-nodemanager:jar:3.2.0:test
-|  |  |  +- com.sun.jersey:jersey-client:jar:1.19:test
 |  |  |  +- io.dropwizard.metrics:metrics-core:jar:4.0.6:test
-|  |  |  +- com.google.inject:guice:jar:4.0:test
-|  |  |  |  +- javax.inject:javax.inject:jar:1:test
-|  |  |  |  \- aopalliance:aopalliance:jar:1.0:test
 |  |  |  \- com.sun.jersey.contribs:jersey-guice:jar:1.19:test
 |  |  +- org.apache.hadoop:hadoop-yarn-server-resourcemanager:jar:3.2.0:test
 |  |  |  \- org.apache.hadoop:hadoop-yarn-server-applicationhistoryservice:jar:3.2.0:test
